requirements.txt
requests==2.32.3
python-dotenv==1.0.1
Pillow>=10.0.0
customtkinter>=5.2.2
